API: Introduced GameService layer
